Implement a GUI window using fyne.io/fyne, alongside go-sdl2
Fyne seems to fit most of what I'd like in a GUI library (easy enough to use, appropriate widgets), but doesn't seem to support natively drawing to a canvas. To get around this we now have two windows: A Fyne GUI with controls to manage the interpreter, and a go-sdl2 window for the actual CHIP-8 display. Will keep playing around with Fyne to see how well it works.